Remove the ``plone.navtree`` content provider
Remove the ``plone.navtree`` content provider that was introduced in the context of the Navigation PLIP (https://github.com/plone/Products.CMFPlone/issues/2516)
The logic has been moved to the viewlet that was invoking it.

Fixes #188
