+++ plone/app/layout/navigation/configure.zcml
@@ -6,14 +6,4 @@
         <allow interface=".interfaces.INavtreeStrategy" />
     </class>
 
-    <!-- a contentprovider -->
-    <adapter
-        factory=".navtree.NavTreeProvider"
-        for="*
-             *
-             *"
-        name="plone.navtree"
-        provides="zope.contentprovider.interfaces.IContentProvider"
-    />
-
 </configure>
