+++ plone/app/layout/viewlets/sections.pt
@@ -21,7 +21,7 @@
         <ul class="plone-nav plone-navbar-nav"
             id="portal-globalnav"
             tal:define="selected_tab python:view.selected_portal_tab">
-          <navtree tal:replace="structure provider:plone.navtree" />
+          <navtree tal:replace="structure view/render_globalnav" />
         </ul>
       </div>
     </div>
